fix(hybrid) broadcast CRUD changes to DPs cluster
This change fixes a bug which manifested itself when Kong:

* It is working on hybrid mode.
* The CP role is being executed by several Kong nodes in a
  cluster.
* A change (any successful CRUD operation) happens in one of these
  nodes.

The desired behaviour is that when this happens, all the CPs in the cluster
would trigger a DP refresh. The bug consists in that only the workers in the
same Kong node as the one which received the CRUD operation do such
refresh.

Example setup: Two Kong nodes in CP role. Node A has workers A1 and A2, and Node
B has workers B1 and B2. The admins use A1's admin API to create a new Service.

Before this fix, A1 would trigger the CP reload, then A2 would "see" the change
(via a worker event) and refresh its DPs. But B1 and B2 would perform such
refresh.

This fix makes A1 trigger a cluster event in addition of the previous
worker-level event. It also makes all workers look for this new cluster event,
and do a DP refresh.

Note that the test accompanying this fix does not exactly replicate the above
setup. There is a concern that the amount of memory this would need (at least 3
kong nodes, with several workers each) would trigger a CI shutdown. Instead, we
reduced the example to the minimum (only A1 and B1 are created) and we check
B1's logs for the reception of the change. The fact that B1 reliably
communicates refreshes its DPs when it receives this notification is assumed,
since the function triggered by the event is the same one used in worker-level
updates, which are already tested.